refactor: refactored lottoFactory to generate both of randomTicket and manualTicket
