+++ src/test/java/lotto/domain/LottoFactoryTest.java
@@ -2,6 +2,9 @@ package lotto.domain;
 
 import org.junit.jupiter.api.Test;
 
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Random;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -9,11 +12,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 class LottoFactoryTest {
 
     @Test
-    void assertNumberOfPurchasedTicket() {
+    void assertNumberOfPurchasedTicketWithManualTicket() {
         int numberOfTickets = 5;
-        Random random = new Random();
 
-        LottoFactory lottoFactory = new LottoFactory(numberOfTickets, random);
-        assertThat(lottoFactory.numberOfPurchasedTicket()).isEqualTo(5);
+        List<String[]> manualTicket = new ArrayList<>();
+        String[] testData = "1, 2, 3, 4, 5, 6".split(", ");
+        manualTicket.add(testData);
+
+        LottoFactory lottoFactory = new LottoFactory(numberOfTickets, manualTicket);
+        assertThat(lottoFactory.numberOfPurchasedTicket()).isEqualTo(6);
     }
 }
